requirements.txt
psycopg2-binary>=2.9
pgvector>=0.2
sentence-transformers>=2.0
ollama>=0.2
python-dotenv>=1.0